requirements.txt
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
wordcloud>=1.9.3
tqdm>=4.66.0
torch>=2.3.0
transformers>=4.42.0
indoNLP>=0.3.4
emosent-py>=0.1.7
Sastrawi>=1.0.1
stopwordsiso>=0.6.1
